GR node (one-file).
