+++ tests/nodes/bragg.cpp
@@ -5,7 +5,7 @@
 #include "classes/partialSet.h"
 #include "classes/speciesSites.h"
 #include "math/rangedVector3.h"
-#include "nodes/gr/gr.h"
+#include "nodes/gr.h"
 #include "nodes/iterableGraph.h"
 #include "nodes/sq/sq.h"
 #include "tests/graphData.h"
